Add files via upload
in seg_noise_filt.m, added section that checks if time series of input signal is long enough to generate meaningful statistics (whether it has enough segments for standard deviation to be reliable). For window size of 60 s, if length of time series is less than 10 mins, then call return statement and return empty array.

in dynam_rho_spikes.m, added section that checks if output from seg_noise_filt.m is empty, if so , call return statement Then, for the section that is responsible for outputting arrays and tables, check if output struct. is empty, if so, do not execute those lines.

This prevents calculating standard deviation and dynamic rho for time series that has less than required number of segments formed, thus removing the possibility of outputting unnecessary data.