Add files via upload
created script for calculating spearman corr. coeff. (\rho) between number of spikes (A) and mean normalized alff (B) for every channel in every run. Additional feature of this script includes the ability to calculate \rho for every channel with A and B in all runs combined. Tables are also generated.