Add files via upload
created scripts to assemble tables of ALFF and ReHo with format given below for every subject, 
col. 1  = event type
col. 2 = channel name
col. 3 = spearman's corr. coeff. btw. ALFF OR ReHo and number of spikes of segments of all runs combined
col. 4 = p-value
col. 5 = Fisher's z
col. 6 = mean of number of spikes across all segments of all runs combined
col. 7 = standard deviation of number of spikes across all segments of all runs combined

Each table is saved under a field which is named after the subject number the table is representing.
For example. to access the table of ALFF for subject 14, use
a.opst.sub14, where a is the variable in which the output file of the script is loaded into, opst is the structure array, and sub14 is the required field